requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
bcrypt==4.1.1
PyJWT==2.8.0
yt-dlp==2024.11.18
gunicorn==21.2.0
certifi>=2024.0.0
boto3==1.34.0
